move operationReturnValues to common as in libsbml
