+++ src/liblx/xml/XMLAttributes.cpp
@@ -47,7 +47,7 @@
 #include <liblx/xml/XMLOutputStream.h>
 /** @endcond */
 #include <liblx/common/util.h>
-#include <liblx/xml/operationReturnValues.h>
+#include <liblx/common/operationReturnValues.h>
 
 /** @cond doxygenIgnored */
 using namespace std;
