+++ src/liblx/xml/XMLNamespaces.cpp
@@ -45,7 +45,7 @@
 //#include <liblx/xml/LXNamespaces.h>
 
 #include <liblx/common/util.h>
-#include <liblx/xml/operationReturnValues.h>
+#include <liblx/common/operationReturnValues.h>
 
 
 /** @cond doxygenIgnored */
