--- src/liblx/xml/operationReturnValues.h
@@ -1,274 +1,0 @@
-/**
- * @file    operationReturnValues.h
- * @brief   Enumeration of values returned by operations within libLX.
- * @author  Sarah Keating
- *
- * <!--------------------------------------------------------------------------
- * This file is part of libLX.  Please visit http://sbml.org for more
- * information about SBML, and the latest version of libLX.
- *
- * Copyright (C) 2019 jointly by the following organizations:
- *     1. California Institute of Technology, Pasadena, CA, USA
- *     2. University of Heidelberg, Heidelberg, Germany
- *
- * Copyright (C) 2013-2018 jointly by the following organizations:
- *     1. California Institute of Technology, Pasadena, CA, USA
- *     2. EMBL European Bioinformatics Institute (EMBL-EBI), Hinxton, UK
- *     3. University of Heidelberg, Heidelberg, Germany
- *
- * Copyright (C) 2009-2013 jointly by the following organizations:
- *     1. California Institute of Technology, Pasadena, CA, USA
- *     2. EMBL European Bioinformatics Institute (EMBL-EBI), Hinxton, UK
- *
- * Copyright (C) 2006-2008 by the California Institute of Technology,
- *     Pasadena, CA, USA
- *
- * Copyright (C) 2002-2005 jointly by the following organizations:
- *     1. California Institute of Technology, Pasadena, CA, USA
- *     2. Japan Science and Technology Agency, Japan
- *
- * This library is free software; you can redistribute it and/or modify it
- * under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation.  A copy of the license agreement is provided
- * in the file named "LICENSE.txt" included with this software distribution and
- * also available online as http://sbml.org/software/libsbml/license.html
- * ---------------------------------------------------------------------- -->*/
-
-#ifndef LIBLX_OPERATION_RETURN_VALUES_H
-#define LIBLX_OPERATION_RETURN_VALUES_H
-
-/**
- * @enum OperationReturnValues_t
- * LibLX diagnostic return codes.
- *
- * Many methods in libLX return a status code to indicate whether the
- * operation requested by the caller succeeded or failed.  This enumeration
- * lists all the possible return codes from any libLX methods.
- */
-typedef enum
-{
-    LIBLX_OPERATION_SUCCESS       = 0
-    /*!< The operation was successful. */
-
-  , LIBLX_INDEX_EXCEEDS_SIZE      = -1
-    /*!< An index parameter exceeded the bounds of a data array or other
-     * collection used in the operation.  This return value is typically
-     * returned by methods that take index numbers to refer to lists
-     * of objects, when the caller has provided an index that exceeds
-     * the bounds of the list.  LibLX provides methods for checking the
-     * size of list/sequence/collection structures, and callers should
-     * verify the sizes before calling methods that take index numbers. */
-
-  , LIBLX_UNEXPECTED_ATTRIBUTE    = -2
-    /*!< The attribute that is the subject of this operation is not valid
-     * for the combination of LX Level and Version for the underlying
-     * object.  This can happen because libLX strives to offer a uniform
-     * API for all LX Levels and Versions, but some object attributes and
-     * elements are not defined for all LX Levels and Versions.  Calling
-     * programs are expected to be aware of which object structures they
-     * are working with, but when errors of this kind occur, they are
-     * reported using this return value. */
-
-  , LIBLX_OPERATION_FAILED        = -3
-    /*!< The requested action could not be performed.  This can occur in
-     * a variety of contexts, such as passing a null object as a parameter
-     * in a situation where it does not make sense to permit a null object.
-     */
-
-  , LIBLX_INVALID_ATTRIBUTE_VALUE = -4
-    /*!< A value passed as an argument to the method is not of a type that
-     * is valid for the operation or kind of object involved.  For example,
-     * this return code is used when a calling program attempts to set an
-     * LX object identifier to a string whose syntax does not conform to
-     * the LX identifier syntax. */
-
-  , LIBLX_INVALID_OBJECT          = -5
-    /*!< The object passed as an argument to the method is not of a type
-     * that is valid for the operation or kind of object involved.  For
-     * example, handing an invalidly-constructed ASTNode to a method
-     * expecting an ASTNode will result in this error. */
-
-  , LIBLX_DUPLICATE_OBJECT_ID     = -6
-    /*!< There already exists an object with this identifier in the
-     * context where this operation is being attempted.  This error is
-     * typically returned in situations where LX object identifiers must be
-     * unique, such as attempting to add two species with the same identifier
-     * to a model. */
-
-  , LIBLX_LEVEL_MISMATCH          = -7
-    /*!< The LX Level associated with the object does not match the Level
-     * of the parent object.  This error can happen when an LX component
-     * such as a species or compartment object is created outside of a model
-     * and a calling program then attempts to add the object to a model that
-     * has a different LX Level defined. */
-
-  , LIBLX_VERSION_MISMATCH        = -8
-    /*!< The LX Version within the LX Level associated with the object
-     * does not match the Version of the parent object.  This error can
-     * happen when an LX component such as a species or compartment object
-     * is created outside of a model and a calling program then attempts to
-     * add the object to a model that has a different LX Level+Version
-     * combination. */
-
-  , LIBLX_INVALID_XML_OPERATION   = -9
-    /*!< The XML operation attempted is not valid for the object or context
-     * involved.  This error is typically returned by the XML interface layer
-     * of libLX, when a calling program attempts to construct or manipulate
-     * XML in an invalid way.  */
-
-  , LIBLX_NAMESPACES_MISMATCH   = -10
-    /*!< The LX Namespaces associated with the object
-     * do not match the LX Namespaces of the parent object.  This error can
-     * happen when an LX component such as a species or compartment object
-     * is created outside of a model and a calling program then attempts to
-     * add the object to a model that has a different LX Namespaces
-     * combination. */
-
-  , LIBLX_DUPLICATE_ANNOTATION_NS   = -11
-    /*!< There already exists a top level annotation with the same namespace as
-     * annotation being appended.  This error is
-     * typically returned in situations where the appendAnnotation function
-     * is being used to add an annotation that has a namespace that is already
-     * present in the existing annotation. */
-
-  , LIBLX_ANNOTATION_NAME_NOT_FOUND   = -12
-    /*!< The existing annotation does not have a top-level element with
-     * the given name. This error is
-     * typically returned in situations where the
-     * replaceTopLevelAnnotationElement function or
-     * the removeTopLevelAnnotationElement function
-     * is being used to replace or remove an annotation with a name that does
-     * not match the name of any top-level element that is already
-     * present in the existing annotation. */
-
-   , LIBLX_ANNOTATION_NS_NOT_FOUND   = -13
-    /*!< The existing annotation does not have a top-level element with
-     * the given namespace. This error is
-     * typically returned in situations where the
-     * replaceTopLevelAnnotationElement function or
-     * the removeTopLevelAnnotationElement function
-     * is being used to replace or remove an annotation with a namespace that does
-     * not match the namespace of any top-level element that is already
-     * present in the existing annotation. */
-
-     , LIBLX_MISSING_METAID   = -14
-    /*!< The requested action cannot be performed as the target object does not have
-     * the metaid attribute set. */
-
-     , LIBLX_DEPRECATED_ATTRIBUTE   = -15
-    /*!< The attribute that is the subject of this operation has been deprecated
-     * for the combination of LX Level and Version for the underlying
-     * object. */
-
-     , LIBLX_USE_ID_ATTRIBUTE_FUNCTION   = -16
-    /*!< For L3V2 use the IdAttribute functions. */
-
-   /* ---------------------------------------------------------------------------
-    *
-    *              Return codes for package extension
-    *
-    * -------------------------------------------------------------------------- */
-
-  , LIBLX_PKG_VERSION_MISMATCH  = -20
-    /*!< The Version of package extension within the LX Level and version
-     * associated with the object does not match the Version of the parent
-     * object. This error can happen when an LX component such as a layout
-     * or group object is created outside of a model and a calling program
-     * then attempts to add the object to a model that has a different LX
-     * Level+Version+Package Version combination. */
-
-   , LIBLX_PKG_UNKNOWN           = -21
-    /*!< The required package extension is unknown. This error is typically
-     * returned when creating an object of SBase derived class with the required
-     * package, creating an object of LXNamespaces or its derived class with the
-     * required package, or invoking functions depending on the required package.
-     * To avoid this error, the library of the required package needs to be
-     * linked.
-     */
-
-   , LIBLX_PKG_UNKNOWN_VERSION    = -22
-    /*!< The required version of the package extension is unknown. This error
-     * is typically returned when creating an object of SBase derived class with
-     * the required package, creating an object of LXNamespaces or its derived
-     * class with the required package, or invoking functions depending on the
-     * required package. This error may be avoided by updating the library of the
-     * required package to be linked.
-     */
-
-   , LIBLX_PKG_DISABLED            = -23
-    /*!< The required package extension is disabled. This error is typically
-     * returned when creating an object of SBase derived class with the required
-     * package, creating an object of LXNamespaces or its derived class with the
-     * required package, or invoking functions depending on the required  package.
-     * To avoid this error, the library of the required package needs to be enabled.
-     */
-
-   , LIBLX_PKG_CONFLICTED_VERSION  = -24
-    /*!< Another version of the required package extension has already been enabled
-     * in the target SBase object, or enabled in the model to/in which the target
-     * object to be added/contained. This error is typically returned  when adding an
-     * object of some SBase derived class with the required package to other SBase
-     * derived object, or when enabling the required package in the target object.
-     * To avoid this error, the conflict of versions need to be avoided.
-     */
-
-   , LIBLX_PKG_CONFLICT            = -25
-    /*!< Another LX package extension for the same URI has already been registered.
-     * This error is typically returned when adding a LX package extension to the
-     * LXExtensionRegistry. To avoid this error, ensure that LX package
-     * extensions are only registered once.
-     */
-
-   /* ---------------------------------------------------------------------------
-    *
-    *              Return codes for conversion extensions
-    *
-    * -------------------------------------------------------------------------- */
-
-  , LIBLX_CONV_INVALID_TARGET_NAMESPACE  = -30
-    /*!< The target namespace is not a valid LX namespace. while
-     * attempting to convert the LX document using
-     * LXLevelVersionConverter::convert() or related methods, the target
-     * namespace has been found to be invalid or unset.  (The function
-     * LXNamespaces::isValidCombination() may be useful in
-     * detecting this situation and preventing the error.)
-     */
-
-  , LIBLX_CONV_PKG_CONVERSION_NOT_AVAILABLE = -31
-    /*!< Conversions involving packages are not available in the specified
-     * routine. This error is typically returned when calling a converter
-     * that does not have the functionality to deal with LX Level&nbsp;3
-     * packages. To avoid this error, ensure that the requested
-     * ConversionProperties specifies packages.
-     */
-
-  , LIBLX_CONV_INVALID_SRC_DOCUMENT = -32
-    /*!< The document on which conversion is being requested is invalid and
-     * the requested conversion cannot be performed. This error is
-     * typically returned when a conversion routine has been given an
-     * invalid target document or the conversion requires a certain degree
-     * of validity that is not present in the document. To avoid this
-     * error, use the LXDocument::checkConsistency() function to find and
-     * resolve errors before passing the document to a conversion method.
-     */
-
-  , LIBLX_CONV_CONVERSION_NOT_AVAILABLE = -33
-    /*!< Conversion with the given properties is not yet available.
-     */
-
-  , LIBLX_CONV_PKG_CONSIDERED_UNKNOWN = -34
-    /*!< The package that is being stripped is not an enabled
-     * package but considered by libLX to be an unrecognized
-     * package.  This error is typically returned when calling
-     * the StripPackage converter requesting that a package
-     * for which code is not available be stripped. Thus the
-     * containing document will not be altered as the elements
-     * and attributes for this package are stored as unknown package
-     * information and will be written out as such.
-     */
-
-} OperationReturnValues_t;
-
-
-#endif  /*ndef LIBLX_OPERATION_RETURN_VALUES_H */
-
