+++ src/liblx/xml/test/TestXMLError.cpp
@@ -45,7 +45,7 @@
 
 #include <check.h>
 #include <XMLError.h>
-#include <liblx/xml/operationReturnValues.h>
+#include <liblx/common/operationReturnValues.h>
 
 /** @cond doxygenIgnored */
