+++ src/liblx/xml/test/TestXMLNamespaces.c
@@ -37,7 +37,7 @@
 #include <liblx/common/common.h>
 #include <liblx/xml/XMLNamespaces.h>
 
-#include <liblx/xml/operationReturnValues.h>
+#include <liblx/common/operationReturnValues.h>
 
 #include <check.h>
