+++ src/liblx/xml/test/TestXMLToken_newSetters.c
@@ -44,7 +44,7 @@
 #include <liblx/common/common.h>
 #include <liblx/xml/XMLToken.h>
 #include <liblx/xml/XMLTriple.h>
-#include <liblx/xml/operationReturnValues.h>
+#include <liblx/common/operationReturnValues.h>
 
 
 #include <liblx/common/extern.h>
